Map View implemented
The Pins for the sites are not dropping yet. Data is there, but
viewForAnnotation is not being hit yet. Also, the LocationListTableView
needs a segmented control to toggle between map view and list view. One
last thing, annotations should move to details view, as well as a
button on the annotation view to create/add locations to a user
customized tour